Add new variable types and reorganize core
