+++ relentless/core/__init__.py
@@ -1,0 +1,6 @@
+from . import collections
+from .collections import *
+from . import math
+from .math import *
+from . import variable
+from .variable import *
